+++ CHANGELOG.md
@@ -38,6 +38,7 @@ If you like this project and find it useful, please consider giving it a star on
 
 ### Added
 
+- [platform]: Add `closureSlidingGate` device with pedestrian opening support. Thanks Ludovic BOUÉ.
 - [devcontainer]: Bump `Dev Container` to v.2.0.0 with a Node Bun stack.
 
 ### Fixed